backend/requirements.txt
# ---------- Django core ----------
Django==5.2.*
django-environ==0.11.*

# ---------- REST API ----------
djangorestframework==3.15.*
djangorestframework-simplejwt==5.3.*
drf-spectacular==0.27.*
drf-spectacular-sidecar==2024.12.*
django-cors-headers==4.6.*
django-filter==24.3

# ---------- Database ----------
psycopg[binary]==3.2.*
dj-database-url==2.3.*

# ---------- Async / tasks ----------
celery[redis]==5.4.*
redis==5.2.*
# django-celery-beat + django-celery-results entegrasyon fazında eklenecek
# (Django 5.2 uyumlu sürümleri 2.8+ bekleniyor)

# ---------- Storage / media ----------
django-storages[s3]==1.14.*
Pillow==11.0.*

# ---------- Integrations ----------
requests==2.32.*
# iyzipay — ödeme fazında eklenecek

# ---------- Security / utils ----------
argon2-cffi==23.1.*
python-slugify==8.0.*
qrcode==8.0

# ---------- Observability ----------
sentry-sdk[django]==2.19.*

# ---------- Dev / test ----------
pytest==8.3.*
pytest-django==4.9.*
pytest-cov==6.0.*
factory-boy==3.3.*
Faker==33.1.*
ruff==0.8.*
black==24.10.*
mypy==1.13.*
django-stubs==5.1.*
djangorestframework-stubs==3.15.*
